+++ libft/ft_substr.c
@@ -14,24 +14,24 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	size_t	i;
-	size_t	j;
+	int		i;
+	int		j;
 	char	*sub;
 
 	i = 0;
 	j = 0;
 	if (!s)
 		return (NULL);
-	if (start > ft_strlen(s))
+	if ((int)start > ft_strlen(s))
 		return (ft_strdup(""));
-	if (len > ft_strlen((s + start)))
+	if ((int)len > ft_strlen((s + start)))
 		len = ft_strlen((s + start));
 	sub = (char *)malloc(sizeof(char) * len + 1);
 	if (!sub)
 		return (NULL);
 	while (s[i])
 	{
-		if (i >= start && j < len)
+		if (i >= (int)start && j < (int)len)
 			sub[j++] = s[i];
 		i++;
 	}
